08.21.2021
Finally I can click my UI Image. It was  hard to do it because I need to be able to click, and drag it using new input system which is not easy to use. but finally, I manage to build new input System that can apply to any object without additional work. all i need is to simply add "Imouse" interface and make sure it can detect some raycast.